Add compatability for IPython3/4/Jupyter in functions that use widgets
Closes #699
